+++ store/driver.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"strings"
 
-	_ "github.com/lib/pq"      // PostgreSQL driver
-	_ "modernc.org/sqlite"     // SQLite driver
+	_ "github.com/lib/pq"  // PostgreSQL driver
+	_ "modernc.org/sqlite" // SQLite driver
 )
 
 // DBType represents database type